+++ MiamCommon/Source/IGraphicSessionManager.cpp
@@ -29,7 +29,7 @@ IGraphicSessionManager::~IGraphicSessionManager()
 {
     delete multiCanvasComponent;
     
-    for(int i=0 ; i<canvasManagers.size(); i++)
+    for(size_t i=0 ; i<canvasManagers.size(); i++)
         delete canvasManagers[i];
 }
